Move beans into mapping component
They are used for mapping, and are not supposed to be a high-level crosscutting component.
