+++ cloud.foundry.cli/src/main/java/cloud/foundry/cli/operations/AbstractOperations.java
@@ -4,8 +4,8 @@ import org.cloudfoundry.operations.CloudFoundryOperations;
 
 /**
  * The base class of all operation classes. An operation class provides an interface for manipulating a certain aspect
- * of a cloud foundry instance. {@link cloud.foundry.cli.crosscutting.beans.Bean Beans} should be used to provide and
- * retrieve data.
+ * of a cloud foundry instance. {@link cloud.foundry.cli.crosscutting.mapping.beans.Bean Beans} should be used to
+ * provide and retrieve data.
  * @param <T> The concrete operations class that is used in the derived class to communicate to the cloud foundry
  *           instance
  */
